+++ lib/fuzz/run.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 
-N="$(cargo fuzz list | wc -l)"
+LIST="$*"
+[ -n "$LIST" ] || LIST=$(echo $(cargo fuzz list))
+list() { for x in $LIST; do echo $x; done; }
+
+N="$(list | wc -l)"
 i=1
-next() { cargo fuzz list | head -n$i | tail -n1; }
+next() { list | head -n$i | tail -n1; }
 while cargo fuzz run "$(next)" -- -max_total_time=600; do
   i=$(( i % N + 1 ))
 done